+++ src/oci_cli/__init__.py
@@ -17,6 +17,7 @@ from .generated import filestorage_cli  # noqa: F401
 from .generated import identity_cli  # noqa: F401
 from .generated import loadbalancer_cli  # noqa: F401
 from .generated import objectstorage_cli  # noqa: F401
+from .generated import resourcesearch_cli  # noqa: F401
 from .generated import virtualnetwork_cli  # noqa: F401
 
 from . import aliasing  # noqa: F401
@@ -27,6 +28,7 @@ from . import database_cli_extended  # noqa: F401
 from . import dns_cli_extended  # noqa: F401
 from . import identity_cli_extended  # noqa: F401
 from . import objectstorage_cli_extended  # noqa: F401
+from . import resourcesearch_cli_extended  # noqa: F401
 from . import email_cli_extended  # noqa: F401
 from . import file_filters  # noqa: F401
 from . import final_command_processor  # noqa: F401